fix(reporting): collapsed common table expressions into subquery joins as mysql 5 is too old for CTE.
Didn't pick this up as was using Postgres locally. Going forward, will use mysql 5.7 for development. This also removed
any Postgres .env config, and removed Postgres setup from the README. The expectation for dev should be we use
the same database and version as production, esp as we do more custom SQL stuff for reporting purposes.